backend/requirements.txt
google-cloud-bigquery==1.23.1
Flask==1.1.1
pyjwt==1.7.1
cryptography
